提交数据管控-规则配置(已知问题修改)

+++ hrds_K/src/main/java/hrds/k/biz/dm/ruleconfig/RuleConfigAction.java
@@ -37,7 +37,7 @@ public class RuleConfigAction extends BaseAction {
 
     @Method(desc = "添加规则", logicStep = "添加规则")
     @Param(name = "dq_definition", desc = "Dq_definition的实体对象", range = "Dq_definition的实体对象")
-    public void addRule(Dq_definition dq_definition) {
+    public void addDqDefinition(Dq_definition dq_definition) {
         //数据校验
         if (StringUtil.isBlank(dq_definition.getLoad_strategy())) {
             throw new BusinessException("规则加载策略为空!");
@@ -58,11 +58,12 @@ public class RuleConfigAction extends BaseAction {
 
     @Method(desc = "删除规则(编号删除)",
             logicStep = "删除规则(编号删除)")
-    @Param(name = "reg_num", desc = "规则编号", range = "long[]类型,数组")
+    @Param(name = "reg_num", desc = "规则编号", range = "long类型,数组")
     public void deleteDqDefinition(long reg_num) {
         //检查数据
         if (checkRegNumIsExist(reg_num)) {
-            Dbo.execute("delete from " + Dq_definition.TableName + " where user_id=?", reg_num);
+            Dbo.execute("delete from " + Dq_definition.TableName + " where user_id=? and reg_num=?", getUserId(),
+                    reg_num);
         }
     }
 
@@ -113,19 +114,27 @@ public class RuleConfigAction extends BaseAction {
             logicStep = "获取规则信息列表")
     @Param(name = "currPage", desc = "分页当前页", range = "大于0的正整数", valueIfNull = "1")
     @Param(name = "pageSize", desc = "分页查询每页显示条数", range = "大于0的正整数", valueIfNull = "10")
-    @Return(desc = "返回值说明", range = "返回值取值范围")
-    public Map<String, Object> getDqDefinitionInfos(int currPage, int pageSize) {
-        //设置查询sql
-        SqlOperator.Assembler asmSql = SqlOperator.Assembler.newInstance();
-        asmSql.clean();
-        asmSql.addSql("select * from " + Dq_definition.TableName + " where user_id=?").addParam(getUserId());
-        //查询并返回结果
-        Map<String, Object> dqDefinitionMap = new HashMap<>();
-        Page page = new DefaultPageImpl(currPage, pageSize);
-        List<Dq_definition> dqDefinitions = Dbo.queryPagedList(Dq_definition.class, page, asmSql.sql(), asmSql.params());
-        dqDefinitionMap.put("dqDefinitions", dqDefinitions);
-        dqDefinitionMap.put("totalSize", page.getTotalSize());
-        return dqDefinitionMap;
+    @Return(desc = "规则信息列", range = "规则信息列")
+    public List<Map<String, Object>> getDqDefinitionInfos(int currPage, int pageSize) {
+        try (DatabaseWrapper db = new DatabaseWrapper()) {
+            //设置查询sql
+            SqlOperator.Assembler asmSql = SqlOperator.Assembler.newInstance();
+            asmSql.clean();
+            asmSql.addSql("select dql.*,? as job_status from " + Dq_definition.TableName + " dql where user_id=?")
+                    .addParam(Job_Effective_Flag.NO.getCode()).addParam(getUserId());
+            //查询
+            Page page = new DefaultPageImpl(currPage, pageSize);
+            List<Map<String, Object>> dqd_list = Dbo.queryPagedList(page, asmSql.sql(), asmSql.params());
+            //处理查询结果
+            Dq_definition dq_definition = new Dq_definition();
+            for (Map<String, Object> dqd : dqd_list) {
+                dq_definition.setReg_num(dqd.get("reg_num").toString());
+                if (DqcExecution.getEffDepJobs(db, dq_definition).size() > 0) {
+                    dqd.put("job_status", Job_Effective_Flag.YES.getCode());
+                }
+            }
+            return dqd_list;
+        }
     }
 
     @Method(desc = "获取规则信息",
@@ -145,6 +154,8 @@ public class RuleConfigAction extends BaseAction {
     @Method(desc = "获取规则类型数据", logicStep = "获取规则类型数据")
     @Return(desc = "规则类型数据", range = "规则类型数据")
     public List<Dq_rule_def> getDqRuleDef() {
+//        List<Map<String, Object>> dq_rule_def_s = Dbo.queryList("select * from " + Dq_rule_def.TableName);
+//        return listToMap(dq_rule_def_s, "case_type", "case_type_desc");
         return Dbo.queryList(Dq_rule_def.class, "select * from " + Dq_rule_def.TableName);
     }
 
@@ -173,9 +184,10 @@ public class RuleConfigAction extends BaseAction {
         //初始化执行sql
         SqlOperator.Assembler asmSql = SqlOperator.Assembler.newInstance();
         asmSql.clean();
-        asmSql.addSql("SELECT * FROM dq_list where user_id = ?").addParam(getUserId());
-        if (StringUtil.isNotBlank(ruleConfSearchBean.getReg_num().toString())) {
-            asmSql.addSql(" and reg_num = ?").addParam(ruleConfSearchBean.getReg_num());
+        asmSql.addSql("SELECT * FROM " + Dq_definition.TableName + " where user_id = ?").addParam(getUserId());
+        if (StringUtil.isNotBlank(ruleConfSearchBean.getReg_num())) {
+            asmSql.addLikeParam(" cast(reg_num as varchar(10))",
+                    '%' + ruleConfSearchBean.getReg_num() + '%');
         }
         if (StringUtil.isNotBlank(ruleConfSearchBean.getTarget_tab())) {
             asmSql.addLikeParam("target_tab", '%' + ruleConfSearchBean.getTarget_tab() + '%');
@@ -189,8 +201,8 @@ public class RuleConfigAction extends BaseAction {
         if (StringUtil.isNotBlank(ruleConfSearchBean.getRule_src())) {
             asmSql.addLikeParam("rule_src", '%' + ruleConfSearchBean.getRule_src() + '%');
         }
-        if (StringUtil.isNotBlank(ruleConfSearchBean.getCase_type())) {
-            asmSql.addORParam("case_type", ruleConfSearchBean.getCase_type().split(","));
+        if (null != ruleConfSearchBean.getCase_type() && ruleConfSearchBean.getCase_type().length > 0) {
+            asmSql.addORParam("case_type", ruleConfSearchBean.getCase_type());
         }
         try (DatabaseWrapper db = new DatabaseWrapper()) {
             //获取搜索结果
@@ -211,10 +223,27 @@ public class RuleConfigAction extends BaseAction {
             db.commit();
             //初始化返回结果List
             List<Map<String, Object>> search_data_list = new ArrayList<>();
-            //根据调度状态处理检索结果
+            //处理检索结果
             dqd_list.forEach(dqd -> {
+                //根据调度状态处理检索结果
                 Job_Effective_Flag job_flag = Job_Effective_Flag.ofEnumByCode(dqd.get("job_status").toString());
-                if (job_flag == Job_Effective_Flag.ofEnumByCode(ruleConfSearchBean.getJob_status())) {
+                if (null != ruleConfSearchBean.getJob_status()) {
+                    for (String job_status : ruleConfSearchBean.getJob_status()) {
+                        if (job_flag == Job_Effective_Flag.ofEnumByCode(job_status)) {
+                            search_data_list.add(dqd);
+                        }
+                    }
+                }
+                //根据规则类型处理检索结果
+                else if (null != ruleConfSearchBean.getCase_type()) {
+                    for (String c_type : ruleConfSearchBean.getCase_type()) {
+                        if (dqd.get("case_type").toString().equals(c_type)) {
+                            search_data_list.add(dqd);
+                        }
+                    }
+                }
+                //规则类型和调度状态都为空,设置检索到的所有数据
+                else {
                     search_data_list.add(dqd);
                 }
             });
@@ -298,5 +327,14 @@ public class RuleConfigAction extends BaseAction {
                 reg_num).orElseThrow(() -> new BusinessException("检查规则reg_num否存在的SQL错误")) == 1;
     }
 
-
+    @Method(desc = "list转Map", logicStep = "result转Map")
+    @Param(name = "list", desc = "结果list", range = "list类型")
+    @Param(name = "colOfKey", desc = "作为key的字段名", range = "String类型")
+    @Param(name = "colOfValue", desc = "作为value的字段名", range = "String类型")
+    @Return(desc = "返回值说明", range = "返回值取值范围")
+    private Map<String, Object> listToMap(List<Map<String, Object>> list, String colOfKey, String colOfValue) {
+        Map<String, Object> map = new HashMap<>();
+        list.forEach(l -> map.put(l.get(colOfKey).toString(), l.get(colOfValue).toString()));
+        return map;
+    }
 }
